requirements.txt
Flask
Flask-SocketIO
opencv-python-headless
picamera2
RPi.GPIO